final_project/requirements.txt
litellm>=1.50.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.8.0
rich>=13.7.0
pyjwt>=2.9.0
beautifulsoup4>=4.12.0
